Fixed compilation error on Win64.
BUG=

Review URL: https://codereview.chromium.org/11538010

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@13199 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -1937,7 +1937,7 @@ void FreeListCategory::Reset() {
 
 
 intptr_t FreeListCategory::CountFreeListItemsInList(Page* p) {
-  intptr_t sum = 0;
+  int sum = 0;
   FreeListNode* n = top_;
   while (n != NULL) {
     if (Page::FromAddress(n->address()) == p) {
@@ -1951,7 +1951,7 @@ intptr_t FreeListCategory::CountFreeListItemsInList(Page* p) {
 
 
 intptr_t FreeListCategory::EvictFreeListItemsInList(Page* p) {
-  intptr_t sum = 0;
+  int sum = 0;
   FreeListNode** n = &top_;
   while (*n != NULL) {
     if (Page::FromAddress((*n)->address()) == p) {
